app/config.py
import os
from dotenv import load_dotenv

load_dotenv()

BOT_TOKEN = os.getenv("BOT_TOKEN")
DATABASE_URL = os.getenv("DATABASE_URL")
CHANNEL_ID = -1003766316650
if not BOT_TOKEN:
    raise ValueError("BOT_TOKEN topilmadi!")

if not DATABASE_URL:
    raise ValueError("DATABASE_URL topilmadi!")